feat: proper illuminance cache (#364)

+++ engine/core/renderer/render_view.cpp
@@ -60,7 +60,7 @@ void RenderView::init() {
     macrofacet_hashset = bank.create_buffer("Macrofacet Hashset Buffer", BufferUsage::Storage, hashset_size, hashkey_size).expect("failed to create macrofacet hashset buffer.");
     macrofacet_shading_commands = bank.create_buffer("Macrofacet Shading Commands Buffer", BufferUsage::Storage | BufferUsage::Indirect, hashset_size, hashkey_size)
                                       .expect("failed to create macrofacet shading commands buffer.");
-    const uint64_t cache_element_size = hashkey_size + sizeof(uint32_t) * 2ull;
+    const uint64_t cache_element_size = hashkey_size + sizeof(uint64_t) * 2ull;
     const uint64_t cache_size = 10'000'000u;
     macrofacet_illuminance_cache =
         bank.create_buffer("Macrofacet Illuminance Cache Buffer", BufferUsage::Storage, cache_size, cache_element_size).expect("failed to create macrofacet illuminance cache buffer.");
